Resolve warnings: array subscript is of type 'char'

+++ libr/asm/p/asm_x86_tab.c
@@ -140,22 +140,22 @@ typedef enum tokentype_t {
  */
 static x86newTokenType getToken(const char *str, size_t *begin, size_t *end) {
 	// Skip whitespace
-	while (isspace(str[*begin]))
+	while (isspace((int)str[*begin]))
 		++(*begin);
 
 	if (!str[*begin]) {                // null byte
 		*end = *begin;
 		return TT_EOF;
 	}
-	else if (isalpha(str[*begin])) {   // word token
+	else if (isalpha((int)str[*begin])) {   // word token
 		*end = *begin;
-		while (isalnum(str[*end]))
+		while (isalnum((int)str[*end]))
 			++(*end);
 		return TT_WORD;
 	}
-	else if (isdigit(str[*begin])) {   // number token
+	else if (isdigit((int)str[*begin])) {   // number token
 		*end = *begin;
-		while (isalnum(str[*end]))     // accept alphanumeric characters, because hex.
+		while (isalnum((int)str[*end]))     // accept alphanumeric characters, because hex.
 			++(*end);
 		return TT_NUMBER;
 	}
